+++ qoqo/src/lib.rs
@@ -23,7 +23,7 @@ use pyo3::prelude::*;
 
 use pyo3::types::PyDict;
 
-use pyo3::wrap_pymodule;
+use pyo3::{wrap_pyfunction, wrap_pymodule};
 
 pub mod operations;
 
@@ -47,7 +47,7 @@ pub use circuitdag::{convert_into_circuitdag, CircuitDagWrapper};
 /// qoqo version information, used for qoqo import/export checks
 pub const QOQO_VERSION: &str = env!("CARGO_PKG_VERSION");
 
-use roqoqo::{RoqoqoBackendError, RoqoqoError};
+use roqoqo::{operations::AVAILABLE_GATES_HQSLANG, RoqoqoBackendError, RoqoqoError};
 use thiserror::Error;
 
 /// Errors that can occur in qoqo.
@@ -89,6 +89,15 @@ pub enum QoqoBackendError {
     RoqoqoBackendError(#[from] RoqoqoBackendError),
 }
 
+/// List of hqslang of all available gates
+#[pyfunction]
+pub fn available_gates_hqslang() -> Vec<String> {
+    AVAILABLE_GATES_HQSLANG
+        .iter()
+        .map(|&s| String::from(s))
+        .collect::<Vec<String>>()
+}
+
 /// Quantum Operation Quantum Operation (qoqo)
 ///
 /// Yes, we use reduplication.
@@ -105,6 +114,7 @@ pub enum QoqoBackendError {
 ///     measurements
 ///     devices
 ///     noise_models
+///     available_gates_hqslang
 ///
 
 #[pymodule]
@@ -113,6 +123,7 @@ fn qoqo(_py: Python, module: &Bound<PyModule>) -> PyResult<()> {
     module.add_class::<QuantumProgramWrapper>()?;
     #[cfg(feature = "circuitdag")]
     module.add_class::<CircuitDagWrapper>()?;
+    module.add_function(wrap_pyfunction!(available_gates_hqslang, module)?)?;
     let wrapper = wrap_pymodule!(operations::operations);
     module.add_wrapped(wrapper)?;
     let wrapper2 = wrap_pymodule!(measurements::measurements);