+++ qoqo/tests/integration/operations/mod.rs
@@ -37,6 +37,8 @@ use pyo3::prelude::*;
 use qoqo_calculator::CalculatorFloat;
 use qoqo_calculator_pyo3::CalculatorFloatWrapper;
 
+use qoqo::available_gates_hqslang;
+
 // helper function to convert CalculatorFloat into a python object
 pub fn convert_cf_to_pyobject(
     py: Python,
@@ -57,4 +59,11 @@ pub fn convert_cf_to_pyobject(
             .unwrap()
             .to_owned(),
     }
+}
+
+#[test]
+fn test_available_gates() {
+    let names = available_gates_hqslang();
+    assert!(names.contains(&"Hadamard".to_string()));
+    assert!(!names.contains(&"Error".to_string()));
 }