+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 This changelog track changes to the qoqo project starting at version v0.5.0
 
+
 ## 1.15.2
 
 ### Fixed in 1.15.2
@@ -16,6 +17,7 @@ This changelog track changes to the qoqo project starting at version v0.5.0
 * Pinned clap version to use rust 1.70 (clap,clap_builder and clap_derive 4.4, clap_lex 0.6.0).
 * Updated to rust 1.70.
 * Updated to jsonschema 0.20.
+* Added list of available gates names.
 
 ## 1.15.1